requirements.txt
anndata>=0.8.0
colorcet>=2.0.1
cvxopt>=1.2.3
csbdeep>=0.6.3
descartes
dynamo-release>=1.3.0
folium>=0.12.1
geopandas>=0.10.2
gpytorch
ipywidgets>=7.7.1
kneed>=0.7.0
kornia>=0.6.4
leidenalg>=0.10.0
loompy>=3.0.5
mapclassify>=2.4.2
matplotlib<=3.5.3
nbconvert
networkx>=2.6.3
# ngs_tools>=1.6.0
numba>=0.46.0
numpy>=1.18.1
opencv-python>=4.5.4.60
pandana
pandas>=0.25.1
paste-bio>=1.4.0
plotly>=5.1.0
POT>=0.8.1
pynndescent>=0.4.8
pysal>=1.14.4
pyro-ppl
python-igraph>=0.7.1
pyvista>=0.39.1
scipy>=1.0
scikit-image>=0.18.0
scikit-learn>=0.19.1
seaborn>=0.9.0
setuptools>=58.0.4
Shapely>=1.8.0
statsmodels>=0.9.0
tensorflow
tqdm>=4.62.3
torch
#torch_scatter>=2.1.2
#torch_sparse>=0.6.18
trame>=2.2.5
trame-server>=2.7.0
typing_extensions>=4.0.1
umap-learn>=0.5.1
vtk>=9.2.2
